+++ packages/backend/src/services/services.ts
@@ -4,12 +4,18 @@ import { UserService } from './UserService';
 import { OrganizationService } from './OrganizationService';
 import { OrganizationModel } from '../models/OrganizationModel';
 import { UserModel } from '../models/UserModel';
+import { SessionModel } from '../models/SessionModel';
 
 const inviteLinkModel = new InviteLinkModel(database);
 const organizationModel = new OrganizationModel(database);
 const userModel = new UserModel(database);
+const sessionModel = new SessionModel(database);
 
-export const userService = new UserService({ inviteLinkModel, userModel });
+export const userService = new UserService({
+    inviteLinkModel,
+    userModel,
+    sessionModel,
+});
 export const organizationService = new OrganizationService({
     organizationModel,
     userModel,